- Merge RobustFederationDescriptionReader class with FederationDescriptionReader - Merge RobustFederationDescriptionReaderTest class with FederationDescriptionReaderTest
